[MOV] l10n_es_edi_facturae{_invoice_period}: merge modules
The invoice period feature for the Spanish Facturae format was added in stable in a patch module. In master, the patch module can be merged into the main one.

task-3704472
