+++ addons/l10n_es_edi_facturae/i18n/l10n_es_edi_facturae.pot
@@ -4,10 +4,10 @@
 #
 msgid ""
 msgstr ""
-"Project-Id-Version: Odoo Server 17.1alpha1+e\n"
+"Project-Id-Version: Odoo Server 17.2alpha1+e\n"
 "Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2024-01-02 13:00+0000\n"
-"PO-Revision-Date: 2024-01-02 13:00+0000\n"
+"POT-Creation-Date: 2024-01-30 09:04+0000\n"
+"PO-Revision-Date: 2024-01-30 09:04+0000\n"
 "Last-Translator: \n"
 "Language-Team: \n"
 "MIME-Version: 1.0\n"
@@ -289,6 +289,11 @@ msgid "FACe Center"
 msgstr ""
 
 #. module: l10n_es_edi_facturae
+#: model_terms:ir.ui.view,arch_db:l10n_es_edi_facturae.view_move_form
+msgid "Factura-e"
+msgstr ""
+
+#. module: l10n_es_edi_facturae
 #: model:ir.model.fields,field_description:l10n_es_edi_facturae.field_account_bank_statement_line__l10n_es_edi_facturae_xml_id
 #: model:ir.model.fields,field_description:l10n_es_edi_facturae.field_account_move__l10n_es_edi_facturae_xml_id
 #: model:ir.model.fields,field_description:l10n_es_edi_facturae.field_account_payment__l10n_es_edi_facturae_xml_id
@@ -364,6 +369,20 @@ msgstr ""
 #. module: l10n_es_edi_facturae
 #: model:ir.model.fields.selection,name:l10n_es_edi_facturae.selection__account_move__l10n_es_edi_facturae_reason_code__14
 msgid "Invoice Class"
+msgstr ""
+
+#. module: l10n_es_edi_facturae
+#: model:ir.model.fields,field_description:l10n_es_edi_facturae.field_account_bank_statement_line__l10n_es_invoicing_period_end_date
+#: model:ir.model.fields,field_description:l10n_es_edi_facturae.field_account_move__l10n_es_invoicing_period_end_date
+#: model:ir.model.fields,field_description:l10n_es_edi_facturae.field_account_payment__l10n_es_invoicing_period_end_date
+msgid "Invoice Period End Date"
+msgstr ""
+
+#. module: l10n_es_edi_facturae
+#: model:ir.model.fields,field_description:l10n_es_edi_facturae.field_account_bank_statement_line__l10n_es_invoicing_period_start_date
+#: model:ir.model.fields,field_description:l10n_es_edi_facturae.field_account_move__l10n_es_invoicing_period_start_date
+#: model:ir.model.fields,field_description:l10n_es_edi_facturae.field_account_payment__l10n_es_invoicing_period_start_date
+msgid "Invoice Period Start Date"
 msgstr ""
 
 #. module: l10n_es_edi_facturae
