+++ addons/l10n_es_edi_facturae/tests/data/expected_invoice_period_document.xml
