--- addons/l10n_es_edi_facturae_invoice_period/i18n/es.po
@@ -1,41 +1,0 @@
-# Translation of Odoo Server.
-# This file contains the translation of the following modules:
-# 	* l10n_es_edi_facturae_invoice_period
-#
-msgid ""
-msgstr ""
-"Project-Id-Version: Odoo Server 17.0+e\n"
-"Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2024-01-23 13:35+0000\n"
-"PO-Revision-Date: 2024-01-23 13:35+0000\n"
-"Last-Translator: \n"
-"Language-Team: \n"
-"Language: es\n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: \n"
-"Plural-Forms: \n"
-
-#. module: l10n_es_edi_facturae_invoice_period
-#: model_terms:ir.ui.view,arch_db:l10n_es_edi_facturae_invoice_period.view_move_form
-msgid "Factura-e"
-msgstr ""
-
-#. module: l10n_es_edi_facturae_invoice_period
-#: model:ir.model.fields,field_description:l10n_es_edi_facturae_invoice_period.field_account_bank_statement_line__l10n_es_invoicing_period_end_date
-#: model:ir.model.fields,field_description:l10n_es_edi_facturae_invoice_period.field_account_move__l10n_es_invoicing_period_end_date
-#: model:ir.model.fields,field_description:l10n_es_edi_facturae_invoice_period.field_account_payment__l10n_es_invoicing_period_end_date
-msgid "Invoice Period End Date"
-msgstr "Fecha final del periodo de facturación"
-
-#. module: l10n_es_edi_facturae_invoice_period
-#: model:ir.model.fields,field_description:l10n_es_edi_facturae_invoice_period.field_account_bank_statement_line__l10n_es_invoicing_period_start_date
-#: model:ir.model.fields,field_description:l10n_es_edi_facturae_invoice_period.field_account_move__l10n_es_invoicing_period_start_date
-#: model:ir.model.fields,field_description:l10n_es_edi_facturae_invoice_period.field_account_payment__l10n_es_invoicing_period_start_date
-msgid "Invoice Period Start Date"
-msgstr "Fecha de inicio del periodo de facturación"
-
-#. module: l10n_es_edi_facturae_invoice_period
-#: model:ir.model,name:l10n_es_edi_facturae_invoice_period.model_account_move
-msgid "Journal Entry"
-msgstr "Asiento contable"
