--- addons/l10n_es_edi_facturae_invoice_period/models/__init__.py
@@ -1,1 +1,0 @@
-from . import account_move
